chore: 移动 HashKindExtensions 到 Enums 命名空间
